fix: split routes
