+++ lib/routes/darwinawards/index.ts
@@ -4,14 +4,14 @@ import got from '@/utils/got';
 import { load } from 'cheerio';
 
 export const route: Route = {
-    path: ['/all', '/'],
+    name: 'Award Winners',
+    example: '/darwinawards',
+    path: '/',
     radar: [
         {
             source: ['darwinawards.com/darwin', 'darwinawards.com/'],
-            target: '',
         },
     ],
-    name: 'Unknown',
     maintainers: ['zoenglinghou', 'nczitzk'],
     handler,
     url: 'darwinawards.com/darwin',
