+++ lib/routes/darwinawards/namespace.ts
@@ -3,4 +3,5 @@ import type { Namespace } from '@/types';
 export const namespace: Namespace = {
     name: 'Darwin Awards',
     url: 'darwinawards.com',
+    categories: ['other'],
 };
