+++ lib/routes/farmatters/namespace.ts
@@ -3,4 +3,5 @@ import type { Namespace } from '@/types';
 export const namespace: Namespace = {
     name: 'Farmatters',
     url: 'farmatters.com',
+    categories: ['new-media'],
 };
