+++ lib/routes/houxu/index.ts
@@ -8,15 +8,15 @@ import { art } from '@/utils/render';
 import path from 'node:path';
 
 export const route: Route = {
-    path: ['/featured', '/index', '/'],
+    name: '热点',
+    maintainers: ['nczitzk'],
+    example: '/houxu',
+    path: '/',
     radar: [
         {
             source: ['houxu.app/'],
-            target: '',
         },
     ],
-    name: 'Unknown',
-    maintainers: [],
     handler,
     url: 'houxu.app/',
 };
