+++ lib/routes/nber/namespace.ts
@@ -3,4 +3,5 @@ import type { Namespace } from '@/types';
 export const namespace: Namespace = {
     name: 'National Bureau of Economic Research',
     url: 'nber.org',
+    categories: ['journal'],
 };
