+++ lib/routes/panewslab/news.ts
@@ -3,18 +3,9 @@ import got from '@/utils/got';
 import { parseDate } from '@/utils/parse-date';
 
 export const route: Route = {
-    path: ['/news', '/newsflash'],
+    path: '/news',
     categories: ['new-media'],
     example: '/panewslab/news',
-    parameters: {},
-    features: {
-        requireConfig: false,
-        requirePuppeteer: false,
-        antiCrawler: false,
-        supportBT: false,
-        supportPodcast: false,
-        supportScihub: false,
-    },
     radar: [
         {
             source: ['panewslab.com/'],
